some tests and moving functionality over to system_validation
